+++ pkg/webscan/ScanDnsSimple.go
@@ -2,19 +2,19 @@ package webscan
 
 import "fmt"
 
-func (engine Engine) ScanDnsSimple() (Engine, error) {
+func (engine Engine) ScanDnsSimple(inputUrl string) (Engine, error) {
 	var (
 		err error
 	)
 
-	fmt.Println("Scanning DNS (simple)...")
+	fmt.Println("Scanning DNS (simple) of", inputUrl, "...")
 
-	engine.DnsScanEngine, err = engine.DnsScanEngine.GetARecords(engine.url)
+	engine.DnsScanEngine, err = engine.DnsScanEngine.GetARecords(inputUrl)
 	if err != nil {
 		return engine, err
 	}
 
-	engine.DnsScanEngine, err = engine.DnsScanEngine.GetAAAARecords(engine.url)
+	engine.DnsScanEngine, err = engine.DnsScanEngine.GetAAAARecords(inputUrl)
 	if err != nil {
 		return engine, err
 	}