+++ pkg/webscan/ScanHttpContent.go
@@ -16,7 +16,7 @@ import (
 
 // TODO return image / media / video / audio / svgs / ... as well
 
-func (engine Engine) ScanHttpContent() (Engine, error) {
+func (engine Engine) ScanHttpContent(inputUrl string) (Engine, error) {
 	var (
 		err error
 
@@ -100,7 +100,7 @@ func (engine Engine) ScanHttpContent() (Engine, error) {
 		}
 
 		if parsedUrl.Host == "" { // Doesn't include hostname
-			parsedUrl.Host = engine.url // Add hostname
+			parsedUrl.Host = inputUrl // Add hostname
 		}
 
 		if !filepath.IsAbs(parsedUrl.Path) { // If not leading '/' in path
@@ -137,7 +137,7 @@ func (engine Engine) ScanHttpContent() (Engine, error) {
 		}
 
 		if parsedUrl.Host == "" { // Doesn't include hostname
-			parsedUrl.Host = engine.url // Add hostname
+			parsedUrl.Host = inputUrl // Add hostname
 		}
 
 		if !filepath.IsAbs(parsedUrl.Path) { // If not leading '/' in path