+++ pkg/webscan/ScanMailConfig.go
@@ -5,13 +5,13 @@ import (
 	"fmt"
 )
 
-func (engine Engine) ScanMailConfig() (Engine, error) {
+func (engine Engine) ScanMailConfig(inputUrl string) (Engine, error) {
 
 	fmt.Println("Scanning mail config...")
 
 	if engine.SubdomainScan {
 		if engine.DkimSelector != "" {
-			engine.mailConfigRecommendations = engine.DnsScanEngine.CheckMailSecurity(engine.url, engine.DkimSelector)
+			engine.mailConfigRecommendations = engine.DnsScanEngine.CheckMailSecurity(inputUrl, engine.DkimSelector)
 		} else {
 			return engine, errors.New("DKIM selector required")
 		}