+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ##
 - Upgrade go version from 1.20 to 1.21
 - `webscan version` now prints the currently used version
+- inputUrl is not stored in webScan.Engine any more, but has to be passed as argument to the Scan functions. It's stored in the Result fields instead.
 
 ## v0.2.1
 - `webscan` now displays status messages during scans